refactor CompileTemplate & FileAutoChange
